Add delimited lexing
This used the stateful lexer machinery and allowed us to force the lexer
to consider each part of the input to have ended at a given point. For
now I'm cheating and dropping delimiters just like whitespace. In
future, if we need them to inform parsing decisions, we can explicitly
allow them in the grammar.