Add error token to output
This should make it easier to process lex errors from outside the
parser. It seems that Nearley doesn't report the information from a
lex error in a machine-readable format. These tokens should cause
parse errors, which will report the token, allowing my code to figure
out what to do with the error.