Replace hard-coded list of stores with list from Redux
This also introduces code to handle unknown stores: if an item's
preferred store is no longer available, it will be assigned to the
default store for grouping purposes. Because this only happens at
display time, re-enabling its preferred store will cause it to pop
back into that list. If it's dragged to another store's list, however,
it will set the preferred store to that store as usual.

The driving factor behind this decision was the desire to ensure that
the user only sees lists for the stores they've defined. This should
avoid confusion caused by removing a store and then seeing items still
in its list. I considered alternative solutions where the store could
only be removed if there were no items in it, but decided that would
be a bother (because you have to manually remove all the items before
deleting the store).