Add Dockerfile
This builds an image which can be used to run the Makefile targets in
a consistent manner. It's got the major advantage that all the context
is pulled across before we do any of the build, which means as soon as
the build context is sent, we can go back to work. Also, it's built in
a consistent environment with all the benefits that brings, so it
should be that much more durable long-term.