Handle getting updated `triggerSend` function into `useEffect`
It's tricky, because the `triggerSend` function needs to know the
`targetId`, which can change over the lifetime of the `useEffect` that
we want to run it from. So we need to ensure we always have the
current version available. The slightly messy way I've chosen to do
this for now is to write the current version to a ref on every
rerender, which then means that the `useEffect` can call whatever's in
that ref at any given time to trigger the current version of
`triggerSend`.