Viewer: Extract and support updating ingredients via dropdown
This required tweaking how we store the text which the Prosemirror
instance is displaying. It's now stored in the Editor component, in
order to allow it to be updated by changes to the dropdown values,
while not getting updated from typing in the editor (we still can't
afford to use a fully-controlled component, because we don't want to
try to store all the Prosemirror state in React yet. I might move to a
different model for wrapping a Prosemirror in React in the future,
which would probably give me all the things I want for free, more or
less, but with what I have, this seems like a reasonable system.

I can get away with completely regenerating the Prosemirror on
dropdown change because the focus should have left the editor, so I
don't think maintaining the cursor position is as important at that
point.

This doesn't yet offer _sensible_ alternative ingredients, mind
you. It just shows that it's possible to update an ingredient via a
dropdown and have the list of ingredients update appropriately.